% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimation.R
\name{base_nw.camp}
\alias{base_nw.camp}
\title{Make Base Population}
\usage{
base_nw.camp(nwstats)
}
\arguments{
\item{nwstats}{Output from \code{\link{calc_nwstats.hiv}}.}
}
\description{
description
}
\details{
This function ...
}

