% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mod.cd4.R
\name{cd4.hiv}
\alias{cd4.hiv}
\title{CD4 Progression Module}
\usage{
cd4.hiv(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function for simulating progression of CD4 in natural
             disease dynamics and in the presence of ART.
}

