% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.circ.R
\name{circ}
\alias{circ}
\title{Circumcision Module}
\usage{
circ(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module for simulating circumcision status at birth/entry into
             the population.
}

