% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/params.R
\name{control.hiv}
\alias{control.hiv}
\title{Control Settings for Stochastic Network Model of HIV-1 Infection in
       Sub-Saharan Africa}
\usage{
control.hiv(simno = 1, nsteps = 100, start = 1, nsims = 1, ncores = 1,
  par.type = "single", initialize.FUN = initialize.hiv,
  aging.FUN = aging.hiv, cd4.FUN = cd4.hiv, vl.FUN = vl.hiv,
  dx.FUN = dx.hiv, tx.FUN = tx.hiv, deaths.FUN = deaths.hiv,
  births.FUN = births.hiv, edges_correct.FUN = edges_correct.hiv,
  resim_nets.FUN = simnet.hiv, infection.FUN = infect.hiv,
  get_prev.FUN = prevalence.hiv, verbose.FUN = verbose.hiv, resim.int = 1,
  clin.array = FALSE, delete.nodes = TRUE, save.int = NULL,
  keep.cpdata = TRUE, save.transmat = FALSE, save.network = TRUE,
  save.nwstats = TRUE, save.other = "attr", verbose = TRUE,
  verbose.int = 100, nwstats.formula = ~edges + meandeg + absdiffby("age",
  "male", 5.38) + degree(0:4, by = "male") + concurrent(by = "male") +
  nodematch("male"), skip.check = TRUE, ...)
}
\arguments{
\item{simno}{Simulation ID number.}

\item{nsteps}{Number of time steps to simulate the model over in whatever unit
implied by \code{time.unit}.}

\item{start}{Starting time step for simulation}

\item{nsims}{Number of simulations.}

\item{ncores}{Number of parallel cores to use for simulation jobs, if using
the \code{EpiModel.hpc} package.}

\item{par.type}{Parallelization type, either of \code{"single"} for multi-core
or \code{"mpi"} for multi-node MPI threads.}

\item{initialize.FUN}{Module to initialize the model at time 1.}

\item{aging.FUN}{Module to age active nodes.}

\item{cd4.FUN}{CD4 progression module.}

\item{vl.FUN}{HIV viral load progression module.}

\item{dx.FUN}{HIV diagnosis module.}

\item{tx.FUN}{HIV treatment module.}

\item{deaths.FUN}{Module to simulate death or exit.}

\item{births.FUN}{Module to simulate births or entries.}

\item{edges_correct.FUN}{Module to adjust the edges coefficient in response
to changes to the population size, with the default function of
\code{\link{edges_correct.hiv}} that preserves mean degree.}

\item{resim_nets.FUN}{Module to resimulate the network at each time step.}

\item{infection.FUN}{Module to simulate disease infection.}

\item{get_prev.FUN}{Module to calculate disease prevalence at each time step,
with the default function of \code{\link{prevalence.hiv}}.}

\item{verbose.FUN}{Module to print simulation progress to screen, with the
default function of \code{\link{verbose.hiv}}.}

\item{resim.int}{Interval unit for resimulation of network, relative to
\code{time.unit} in \code{\link{param.hiv}}.}

\item{clin.array}{If \code{TRUE}, save an array of individual-level attribute
history for treatment status, viral load, and CD4 level.}

\item{delete.nodes}{If \code{TRUE}, delete vertices upon exit/death, otherwise
deactivate them.}

\item{save.int}{Interval of time steps at which dat object should be saved to
disk.}

\item{keep.cpdata}{Keep checkpoint interval data.}

\item{save.transmat}{Save out transmission matrix.}

\item{save.network}{Save out network objects (in \code{simOut} function).}

\item{save.nwstats}{Save out network statistics.}

\item{save.other}{Other list elements of dat to save out.}

\item{verbose}{If \code{TRUE}, print progress to console.}

\item{verbose.int}{Interval for printing progress to console.}

\item{nwstats.formula}{Formula for network statistics on the network.}

\item{skip.check}{If \code{TRUE}, skips the error check for parameter values,
initial conditions, and control settings before running the models.
This is suggested only if encountering unnecessary errors when running
new models.}

\item{...}{Additional arguments passed to the function.}
}
\description{
This function sets the control settings for the stochastic
             network models in the \code{epimethods} package.
}
\details{
This function sets the parameters for the models.
}

