% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mod.deaths.R
\name{deaths.hiv}
\alias{deaths.hiv}
\title{Deaths Module}
\usage{
deaths.hiv(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module for simulating deaths among susceptible and infected
             persons within the population.
}

