% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EpiModel.hiv-package.R
\docType{package}
\name{EpiModel.hiv-package}
\alias{EpiModel.hiv-package}
\title{EpiModel.hiv Package}
\description{
\tabular{ll}{
   Package: \tab EpiModel.hiv\cr
   Type: \tab Package\cr
   Version: \tab 0.1\cr
   Date: \tab 2015-04-15\cr
   License: \tab GPL (>= 2)\cr
   LazyLoad: \tab yes\cr
}
}
\details{
EpiModel extensions for modeling HIV infection among heterosexual populations.
}
\keyword{package}

