% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelHIVhet-package.R
\docType{package}
\name{EpiModelHIVhet-package}
\alias{EpiModelHIVhet-package}
\title{EpiModelHIVhet Package}
\description{
\tabular{ll}{
   Package: \tab EpiModelHIVhet\cr
   Type: \tab Package\cr
   Version: \tab 0.5\cr
   Date: \tab 2016-06-21\cr
   License: \tab GPL (>= 3)\cr
   LazyLoad: \tab yes\cr
}
}
\details{
EpiModel extensions for modeling HIV infection among heterosexual populations.
}
\keyword{package}

