% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.init.R
\name{initialize.hiv}
\alias{initialize.hiv}
\title{Initialization Module}
\usage{
initialize.hiv(x, param, init, control, s)
}
\arguments{
\item{x}{An \code{EpiModel} object of class \code{\link{netest}}.}

\item{param}{An \code{EpiModel} object of class \code{\link{param.hiv}}.}

\item{init}{An \code{EpiModel} object of class \code{\link{init.hiv}}.}

\item{control}{An \code{EpiModel} object of class \code{\link{control.hiv}}.}

\item{s}{Simulation number, used for restarting dependent simulations.}
}
\value{
This function returns the updated \code{dat} object with the initialized values
for demographics and disease-related variables.
}
\description{
This function initializes the master \code{dat} object on which
             data are stored, simulates the initial state of the network, and
             simulates disease status and other attributes.
}

