% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mod.tx.R
\name{tx.hiv}
\alias{tx.hiv}
\title{HIV Anti-Retroviral Treatment Module}
\usage{
tx.hiv(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function for simulating HIV therapy after diagnosis,
             including adherence and non-adherence to ART.
}

