% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mod.updatenwp.R
\name{update_nwp}
\alias{update_nwp}
\title{Update Network Parameters Module}
\usage{
update_nwp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function to update network parameters (change statistics
             and vertex attribute vectors) for simulate_network function.
}

